Improve Memory Usage Finder to switch when set up services
