+++ src/MemStat.Net/IMemoryUsageFinder.cs
@@ -3,6 +3,7 @@ namespace MemStat.Net;
 
 public interface IMemoryUsageFinder
 {
+    public ResultBox<object> GetRawMemoryUsageObject();
     public ResultBox<UnitValue> ReceiveCurrentMemoryUsage();
     public ResultBox<double> GetTotalMemoryUsage();
     public ResultBox<double> GetMemoryUsagePercentage();
