+++ src/MemStat.Net/MemoryUsageFinder.cs
@@ -4,115 +4,31 @@ namespace MemStat.Net;
 
 public class MemoryUsageFinder : IMemoryUsageFinder
 {
-    public OptionalValue<MacVmStat> MacVmStat { get; private set; } = OptionalValue<MacVmStat>.Empty;
-    public OptionalValue<LinuxMemoryInfo> LinuxMemoryInfo { get; private set; } = OptionalValue<LinuxMemoryInfo>.Empty;
-    public OptionalValue<WindowsMemoryInfo.MEMORYSTATUSEX> WindowsComputerInfo { get; private set; }
-        = OptionalValue<WindowsMemoryInfo.MEMORYSTATUSEX>.Empty;
+    public OptionalValue<IMemoryUsageFinder> MemoryUsageFinderInternal { get; private set; }
+        = OptionalValue<IMemoryUsageFinder>.Empty;
+
+    public ResultBox<object> GetRawMemoryUsageObject() => MemoryUsageFinderInternal.Match(
+        finder => finder.GetRawMemoryUsageObject(),
+        () => new InvalidOperationException(
+            "MemoryUsageFinderInternal is not set. Please run ReceiveCurrentMemoryUsage() first."));
     public ResultBox<UnitValue> ReceiveCurrentMemoryUsage() =>
-        ResultBox.Start.Conveyor(
-            _ => _ switch
-            {
-                not null when RuntimeInformation.IsOSPlatform(OSPlatform.Linux) => ReceiveCurrentMemoryUsageLinux(),
-                not null when RuntimeInformation.IsOSPlatform(OSPlatform.OSX) => ReceiveCurrentMemoryUsageOSX(),
-                not null when RuntimeInformation.IsOSPlatform(OSPlatform.Windows) => ReceiveCurrentMemoryUsageWindows(),
-                _ => new PlatformNotSupportedException()
-            });
-
-    ResultBox<double> IMemoryUsageFinder.GetTotalMemoryUsage() => GetTotalMemoryUsage();
-    ResultBox<double> IMemoryUsageFinder.GetMemoryUsagePercentage() => GetMemoryUsagePercentage();
-
-    public ResultBox<UnitValue> ReceiveCurrentMemoryUsageOSX() => ProcessToStringList
-        .GetProcessOutput("sh", "-c vm_stat")
-        .Conveyor(lines => ResultBox.WrapTry(() => Net.MacVmStat.Parse(lines, DateTime.UtcNow)))
-        .Do(stat => MacVmStat = stat)
-        .Conveyor(() => ResultBox.UnitValue);
-    public ResultBox<UnitValue> ReceiveCurrentMemoryUsageLinux() => ProcessToStringList
-        .GetProcessOutput("sh", "-c free")
-        .Conveyor(lines => ResultBox.WrapTry(() => Net.LinuxMemoryInfo.Parse(lines, DateTime.UtcNow)))
-        .Do(stat => LinuxMemoryInfo = stat)
-        .Conveyor(() => ResultBox.UnitValue);
-    public ResultBox<UnitValue> ReceiveCurrentMemoryUsageWindows() => ResultBox
-        .WrapTry(() => new WindowsMemoryInfo.MEMORYSTATUSEX())
-        .Combine(info => ResultBox.WrapTry(() => WindowsMemoryInfo.GlobalMemoryStatusEx(info)))
-        .Do(value => WindowsComputerInfo = value.Value1)
-        .Conveyor(() => ResultBox.UnitValue);
-
-    public ResultBox<double> GetTotalMemoryUsage() =>
-        ResultBox.Start.Conveyor(
-            _ => _ switch
-            {
-                not null when RuntimeInformation.IsOSPlatform(OSPlatform.Linux) => GetLinuxTotalMemory(),
-                not null when RuntimeInformation.IsOSPlatform(OSPlatform.OSX) => GetMacTotalMemory(),
-                not null when RuntimeInformation.IsOSPlatform(OSPlatform.Windows) => GetWindowsTotalMemory(),
-                _ => new PlatformNotSupportedException()
-            });
-
-    public ResultBox<double> GetMemoryUsagePercentage() =>
-        ResultBox.Start.Conveyor(
-            _ => _ switch
-            {
-                not null when RuntimeInformation.IsOSPlatform(OSPlatform.Linux) => GetLinuxMemoryUsagePercentage(),
-                not null when RuntimeInformation.IsOSPlatform(OSPlatform.OSX) => GetMacMemoryUsagePercentage(),
-                not null when RuntimeInformation.IsOSPlatform(OSPlatform.Windows) => GetWindowsMemoryUsagePercentage(),
-                _ => new PlatformNotSupportedException()
-            });
-
-    private ResultBox<double> GetMacTotalMemory() => MacVmStat.Match(
-        stat => (Net.MacVmStat.TotalPages(stat) * stat.PageSize).ToResultBox(),
-        () => new InvalidOperationException("VmStat is not set. Please run ReceiveCurrentMemoryUsage() first."));
-
-    private ResultBox<double> GetWindowsTotalMemory() => WindowsComputerInfo.Match(
-        info => ((double)info.ullTotalPhys).ToResultBox(),
-        () => new InvalidOperationException("ComputerInfo is not set. Please run ReceiveCurrentMemoryUsage() first."));
-    private ResultBox<double> GetMacMemoryUsagePercentage() => MacVmStat.Match(
-        stat => Net.MacVmStat.MemoryUsagePercentage(stat).ToResultBox(),
-        () => new InvalidOperationException("VmStat is not set. Please run ReceiveCurrentMemoryUsage() first."));
-
-    private ResultBox<double> GetLinuxTotalMemory() => LinuxMemoryInfo.Match(
-        info => ((double)info.Total).ToResultBox(),
-        () => new InvalidOperationException("MemoryInfo is not set. Please run ReceiveCurrentMemoryUsage() first."));
-
-    private ResultBox<double> GetLinuxMemoryUsagePercentage() => LinuxMemoryInfo.Match(
-        info => Net.LinuxMemoryInfo.MemoryUsagePercentage(info).ToResultBox(),
-        () => new InvalidOperationException("MemoryInfo is not set. Please run ReceiveCurrentMemoryUsage() first."));
-    private ResultBox<double> GetWindowsMemoryUsagePercentage() => WindowsComputerInfo.Match(
-        info => (((double)info.ullTotalPhys - info.ullAvailPhys) / info.ullTotalPhys).ToResultBox(),
-        () => new InvalidOperationException("ComputerInfo is not set. Please run ReceiveCurrentMemoryUsage() first."));
-}
-public class WindowsMemoryInfo
-{
-
-    [DllImport("kernel32.dll", CharSet = CharSet.Auto, SetLastError = true)]
-    [return: MarshalAs(UnmanagedType.Bool)]
-    internal static extern bool GlobalMemoryStatusEx([In] [Out] MEMORYSTATUSEX lpBuffer);
-
-    public static void GetMemoryStatus(out double totalMemoryGB, out double availableMemoryGB)
+        GetEnvironmentalMemoryUsageFinder()
+            .ToResultBox()
+            .Do(finder => MemoryUsageFinderInternal = OptionalValue<IMemoryUsageFinder>.FromValue(finder))
+            .Conveyor(finder => finder.ReceiveCurrentMemoryUsage());
+    public ResultBox<double> GetTotalMemoryUsage() => MemoryUsageFinderInternal.Match(
+        finder => finder.GetTotalMemoryUsage(),
+        () => new InvalidOperationException(
+            "MemoryUsageFinderInternal is not set. Please run ReceiveCurrentMemoryUsage() first."));
+    public ResultBox<double> GetMemoryUsagePercentage() => MemoryUsageFinderInternal.Match(
+        finder => finder.GetMemoryUsagePercentage(),
+        () => new InvalidOperationException(
+            "MemoryUsageFinderInternal is not set. Please run ReceiveCurrentMemoryUsage() first."));
+    public static IMemoryUsageFinder GetEnvironmentalMemoryUsageFinder() => UnitValue.Unit switch
     {
-        var memoryStatus = new MEMORYSTATUSEX();
-        if (GlobalMemoryStatusEx(memoryStatus))
-        {
-            totalMemoryGB = ConvertBytesToGB(memoryStatus.ullTotalPhys);
-            availableMemoryGB = ConvertBytesToGB(memoryStatus.ullAvailPhys);
-        } else
-        {
-            throw new InvalidOperationException("Failed to retrieve memory status.");
-        }
-    }
-    private static double ConvertBytesToGB(ulong bytes) => bytes / (1024.0 * 1024.0 * 1024.0);
-    [StructLayout(LayoutKind.Sequential, CharSet = CharSet.Auto)]
-    public class MEMORYSTATUSEX
-    {
-        // @formatter:off
-        public uint dwLength;
-        public uint dwMemoryLoad;
-        public ulong ullTotalPhys;
-        public ulong ullAvailPhys;
-        public ulong ullTotalPageFile;
-        public ulong ullAvailPageFile;
-        public ulong ullTotalVirtual;
-        public ulong ullAvailVirtual;
-        public ulong ullAvailExtendedVirtual;
-        public MEMORYSTATUSEX() => dwLength = (uint)Marshal.SizeOf(typeof(MEMORYSTATUSEX));
-        // @formatter:on
-    }
+        not null when RuntimeInformation.IsOSPlatform(OSPlatform.Linux) => new LinuxMemoryUsageFinder(),
+        not null when RuntimeInformation.IsOSPlatform(OSPlatform.OSX) => new MacMemoryUsageFinder(),
+        not null when RuntimeInformation.IsOSPlatform(OSPlatform.Windows) => new WindowsMemoryUsageFinder(),
+        _ => throw new PlatformNotSupportedException()
+    };
 }
